refactor: _components로 이동
