+++ src/app/(home)/page.tsx
@@ -1,4 +1,4 @@
-import Test from './components/Test';
+import Test from './_components/Test';
 
 export default function Page() {
   return <Test />;
